separate and simplify the listen/connect iroh-net examples
