+++ iroh/examples/collection-fetch.rs
@@ -2,8 +2,7 @@
 //! Will only work with collections that contain text, and is meant as a companion to the and `collection-provide` example.
 //!
 //! This is using an in memory database and a random node id.
-//! Run the `collection-provide` example, copy the ticket, and run this example from the project root:
-//!     $ cargo run -p collection-fetch [TICKET]
+//! Run the `collection-provide` example, which will give you instructions on how to run this example.
 use anyhow::{bail, Context, Result};
 use iroh::{client::BlobDownloadProgress, rpc_protocol::BlobDownloadRequest};
 use iroh_bytes::BlobFormat;
